+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [🚀] `ScrollableX`增加`offsetToStartEdge`, `offsetToEndEdge`扩展;
 - [🚀] `ScrollableX`增加`animateToNoOverscroll`和`animateByNoOverscroll`扩展;
 - [🚀] `IterableX`增加`removeAfter`和`removeBefore`;
+- [🚀] 增加`CircularBuffer`;
 
 ## 0.50.2
 - [🚀] `SelectableListX`增加`reset`扩展;